closes #62

+++ repository/comment_repo.py
@@ -5,9 +5,10 @@ from messages import Message
 
 def delete_book_comments(book_id,db_session):
 
-
     try:
-        db_session.query(Comment).filter(Comment.book_id == book_id).delete()
+       comment =  db_session.query(Comment).filter(Comment.book_id == book_id).first()
+       if comment is not None:
+           db_session.query(Comment).filter(Comment.book_id == book_id).delete()
 
     except:
         raise Http_error(404,Message.MSG20)
